The exchange code in /python/IPO/banksay.py was missing a cast to integer, this resulted in a traceback on Python 3.x.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/maps@11477 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ trunk/python/IPO/banksay.py
@@ -120,7 +120,7 @@ elif text[0] == 'exchange':
 					# prevents the player from carrying too
 					# many coins.
 					id = activator.Map.CreateObject('platinum coin', x, y)
-					CFItemBroker.Item(id).add(amount*(exchange_rate/50))
+					CFItemBroker.Item(id).add(int(amount*(exchange_rate/50)))
 					activator.Take(id)
 
 					message = random.choice(thanks_message)
